Make pool.alloc and alloc_bulk functions safe to call
The safety is now checked by the packet's phantom reference to the
mempool.

The rx function also gains some lifetime annotations, previously it was
marked as safe, but lifetime of returned packets was unbounded.

This commit also removes a comment about packet being compat with *mbuf,
the repr(transparent) attribute should verify that.